#190: Rename query_filter_factory.py
